+++ src/modules/user/usecase/user.create.ts
@@ -1,7 +1,7 @@
 import User from "../entity/User";
 import UserRepository from "../user.repository";
 import JWTService from "../../application/services/jwt.service";
-import SendEmail from '../usecase/email.send';
+import SendEmail from './email.send';
 
 export default class CreateUser {
