+++ src/modules/user/usecase/user.login.ts
@@ -1,7 +1,7 @@
 import JWTService from "../../application/services/jwt.service";
 import DomainError from "../../application/error/DomainError";
 import PasswordService from "../../application/services/password.service";
-import UserRepository from "../../user/user.repository";
+import UserRepository from "../user.repository";
 
 
 export default class Login {
