+++ src/modules/user/usecase/users.get.ts
@@ -1,4 +1,4 @@
-import UserRepository from "./../user.repository";
+import UserRepository from "../user.repository";
